Refactor to use 'double' instead of 'float' in expr classes
Replaces most Cython 'float' type annotations with 'double' in expression-related classes for improved type consistency and precision. Updates function signatures, class attributes, and internal helper functions accordingly. Also adds TYPE_CHECKING import guard for 'double' type hinting.

+++ src/pyscipopt/scip.pxd
@@ -2200,7 +2200,7 @@ cdef class Expr(UnaryOperatorMixin):
 
     cdef dict _to_dict(self, Expr other, bool copy = *)
 
-    cpdef list _to_node(self, float coef = *, int start = *)
+    cpdef list _to_node(self, double coef = *, int start = *)
 
     cdef Expr copy(self, bool copy = *, object cls = *)
